커맨드 리스트 이벤트로 전달

+++ Server/Server/TCPServer.h
@@ -115,11 +115,11 @@ typedef struct CommandPacket
 	//std::vector<int> commandblockdata_2{ 0 };
 	//std::vector<int> commandblockdata_3{ 0 };
 
-	int commandblockindex[COMMANDS] = {-1,-1,-1,-1,-1,-1,-1,-1,-1};
-	int commandblockdata_0[COMMANDS] = { 0, };
-	int commandblockdata_1[COMMANDS] = { 0, };
-	int commandblockdata_2[COMMANDS] = { 0, };
-	int commandblockdata_3[COMMANDS] = { 0, };
+	float commandblockindex[COMMANDS] = {-1,-1,-1,-1,-1,-1,-1,-1,-1};
+	float commandblockdata_0[COMMANDS] = { 0, };
+	float commandblockdata_1[COMMANDS] = { 0, };
+	float commandblockdata_2[COMMANDS] = { 0, };
+	float commandblockdata_3[COMMANDS] = { 0, };
 	
 
 }CommandPacket;
